# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScenarioInputCompanyType = typeof ScenarioInputCompanyType[keyof typeof ScenarioInputCompanyType];


export const ScenarioInputCompanyType = {
  Public: 'Public',
  Private: 'Private',
  'PE-backed': 'PE-backed',
} as const;

export interface ScorecardMetric {
  name: string;
  weight: number;
}

export interface YearlyCompensation {
  year: string;
  base: number;
  incentive: number;
  cash: number;
  equity: number;
}

export interface ScenarioInput {
  companyType: ScenarioInputCompanyType;
  candidateRole: string;
  year: number;
  /**
     * @minimum 1
     * @maximum 15
     */
  yearsToModel: number;
  /** @minimum 0 */
  marketReference: number;
  /** @minimum 0 */
  baseSalary: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  annualIncreasePercent: number;
  /**
     * @minimum 0
     * @maximum 1000
     */
  targetIncentivePercent: number;
  /** @minimum 0 */
  maximumMultiple: number;
  metrics: ScorecardMetric[];
  yearlyCompensation: YearlyCompensation[];
}

export interface LeadInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  company: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  role: string;
  consent: true;
  scenario: ScenarioInput;
}

export interface LeadReceipt {
  id: number;
  pdfBase64: string;
  fileName: string;
  emailSent: boolean;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  company: string;
  role: string;
  createdAt: string;
  scenario: ScenarioInput;
}