# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScenarioInput } from './scenarioInput';

export interface Lead {
  id: number;
  name: string;
  email: string;
  company: string;
  role: string;
  createdAt: Date;
  scenario: ScenarioInput;
}