# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScenarioInput } from './scenarioInput';

export interface LeadInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @maxLength 320 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  company: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  role: string;
  consent: true;
  scenario: ScenarioInput;
}