# lib/api-zod/src/generated/types/leadReceipt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadReceipt {
  id: number;
  pdfBase64: string;
  fileName: string;
  emailSent: boolean;
}