# lib/api-zod/src/generated/types/scenarioInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScenarioInputCompanyType } from './scenarioInputCompanyType';
import type { ScorecardMetric } from './scorecardMetric';
import type { YearlyCompensation } from './yearlyCompensation';

export interface ScenarioInput {
  companyType: ScenarioInputCompanyType;
  candidateRole: string;
  year: number;
  /**
     * @minimum 1
     * @maximum 15
     */
  yearsToModel: number;
  /** @minimum 0 */
  marketReference: number;
  /** @minimum 0 */
  baseSalary: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  annualIncreasePercent: number;
  /**
     * @minimum 0
     * @maximum 1000
     */
  targetIncentivePercent: number;
  /** @minimum 0 */
  maximumMultiple: number;
  metrics: ScorecardMetric[];
  yearlyCompensation: YearlyCompensation[];
}