# lib/api-zod/src/generated/types/yearlyCompensation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface YearlyCompensation {
  year: string;
  base: number;
  incentive: number;
  cash: number;
  equity: number;
}